+++ src/framework/piral-core/src/types/api.ts
@@ -58,6 +58,10 @@ export interface PageComponentProps<T = any, S = any> extends RouteBaseProps<T, 
    * The meta data registered with the page.
    */
   meta: PiralPageMeta;
+  /**
+   * The children of the page.
+   */
+  children: ReactNode;
 }
 
 /**